% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addchar.R
\name{addchar}
\alias{addchar}
\title{Insert Characters into an Existing String}
\usage{
\code{addchar(string, characters, position)}
}
\arguments{
\item{string}{Character. A string of class \code{character}.}

\item{characters}{Character. A string containing the additional characters to insert.}

\item{position}{Numeric. The position within \code{string} where
\code{characters} should be added.}
}
\description{
Given a string, and a starting position, this function will enlarge
the string to include additional characters.
}
\examples{
# Returns "The quick brown fox jumps over the lazy dog."
orig_string <- "The brown fox jumps over the lazy dog."
new_string  <- "quick "

addchar(orig_string, new_string, 4)

}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
