% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_dtc.R
\name{adjust_dtc}
\alias{adjust_dtc}
\title{Add or Subtract Time from a Date-Time Object}
\usage{
adjust_dtc(time1, timediff, units, tz, POSIX)
}
\arguments{
\item{time1}{Date-time object in character format.}

\item{timediff}{Value by which to increment or decrement.}

\item{units}{Units of \code{timediff}; i.e. minutes, hours, days or weeks.}

\item{tz}{POSIX-compliant time zone.}

\item{POSIX}{Boolean; if \code{TRUE}, output POSIXct object, otherwise output character string.}
}
\value{
Either character string or POSIXct object, depending on \code{POSIX} parameter.
}
\description{
Computes difftime for date-time objects in R, removing separators from CDISC-typical character strings as needed.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
