% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clmest.R
\name{clmest}
\alias{clmest}
\title{Compute Confidence Limits Based on One-Sample t-Test}
\usage{
clmest(datavec, conf.level = 0.95)
}
\arguments{
\item{datavec}{Numeric. Vector of numeric values.}

\item{conf.level}{Numeric. Confidence level required for t-test when estimating limits.}
}
\description{
This function emulates the behaviour of SAS's MEANS procedure, when specifying CLM. This is a convenience
wrapper for the \code{\link{t.test}} function, primarily useful for plotting purposes.
}
\examples{
counts <- c(18,17,15,20,10,20,25,13,12)
clmest(counts)

}
\references{
\url{https://v8doc.sas.com/sashtml/proc/z0608466.htm}
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
