% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_info.R
\name{col_info}
\alias{col_info}
\title{Produce an Object Detail Table}
\usage{
col_info(x)
}
\arguments{
\item{x}{A \code{\link{data.frame}}, \code{\link{data.table}}, or vector.}
}
\value{
\code{data.frame} containing the \code{\link{mode}}, \code{\link{class}}, \code{\link{storage.mode}}, and \code{\link{factor}} status.
}
\description{
This function will print object details to the console for each column in a \code{data.frame}, \code{data.table},
or an individual vector.
}
\details{
Currently, the function will return the \code{\link{mode}}, \code{\link{storage.mode}}, and \code{\link{class}} of
an object. It will also specify whether or not the object is a \link{factor}.

While the \code{\link{storage.mode}} and \code{\link{class}} of an object are often identical, they may not be,
e.g. when a vector has been encoded as a factor.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
