% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{Quickly view a package DESCRIPTION}
\usage{
desc(package)
}
\arguments{
\item{package}{A character string corresponding to a
valid R package name.}
}
\description{
This is a simple convenience function to easily view a
package DESCRIPTION file for a given R package. Since
package maintainers often omit a generic help file
including  _PACKAGE, typing help("somePackage") is not
always successful.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
