% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_all.R
\name{detach_all}
\alias{detach_all}
\title{Detach All Packages}
\usage{
detach_all(verbose=TRUE, unload=TRUE)
}
\arguments{
\item{verbose}{Should a message be printed to the console naming all the detached packages? \code{TRUE} or \code{FALSE}.}

\item{unload}{Should the packages be unloaded from the namespace? \code{TRUE} or \code{FALSE}.}
}
\description{
This function will attempt to detach all packages loaded by the user during the current R session.
}
\details{
It's probably smarter to restart R, but being able to detach and unload all packages can be handy at times.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
