% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_dtc.R
\name{difftime_dtc}
\alias{difftime_dtc}
\title{Time Differences for Date/Time Objects}
\usage{
difftime_dtc(date1, date2, units, tz, as.real)
}
\arguments{
\item{date1}{A date, a character string in date-time format.}

\item{date2}{A second date, a character string in date-time format.}

\item{units}{Units for calculation, defaults to \code{hours}.}

\item{tz}{Time-zone, defaults to "\code{UTC}".}

\item{as.real}{Boolean; should result be returned as a real number?}
}
\value{
Either a numeric value corresponding to the time difference between the inputs, or a \code{difftime} object.
}
\description{
This function takes two date-time objects, with or without date separator,
and estimates the time difference between them.
}
\details{
The function allows for date-time objects of class character or POSIXct/POSIXlt;
it will convert either or both inputs to POSIX-compliant objects as required.

The result is returned as a numerical difference (real number), or as a difftime
object.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
