% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseFreq.R
\name{doseFreq}
\alias{doseFreq}
\title{Compute Dose Frequency for Longitudinal Analyses in Long Format, with Handling for NONMEM-compliant data.}
\usage{
doseFreq(dosedata, dosevar = "DOSE", timevar = "DAY", idvar = "ID",
  nmdata = FALSE)
}
\arguments{
\item{dosedata}{A data frame, including individual-level data organized with a unique key for each subject (\code{idvar}), and a secondary key corresponding to observed time (\code{timevar}). Neither should be missing.}

\item{dosevar}{A character string specifying the column containing dose amount information.}

\item{timevar}{A character string specifying the column containing time information, assumed to be days since start of observation.}

\item{idvar}{A unique, subject-level ID which may identify individual participants in the data set.}

\item{nmdata}{Boolean; if \code{TRUE}, assume data are in NONMEM-like format, and subset to rows containing \code{EVID==1}.}
}
\value{
A data frame, in which rows correspond to unique values of \code{idvar} and columns correspond to \code{qd}, \code{bid}, \code{tid}, and \code{qid} dosing.
}
\description{
This function produces tabular output which may be helpful for comparing dosing frequencies by individual in exploratory analyses.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
