% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2dttm.R
\name{dt2dttm}
\alias{dt2dttm}
\title{Create Date/Time Object}
\usage{
dt2dttm(dateobj, timeobj, format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC")
}
\arguments{
\item{dateobj}{A date formatted as a character string.}

\item{timeobj}{A time formatted as a character string.}

\item{format}{A character string containing the format of the resulting date-time object.}

\item{tz}{The desired \link{timezone} to specify for the result.}
}
\value{
A object of class \code{POSIXct}, formatted according to \code{format} and \code{tz}.
}
\description{
This function takes a time and a date, and melds them together.
}
\examples{
dt2dttm("2018-04-04", "15:15:00") # returns "2018-04-04 15:15:00 UTC"

}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
