% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr_mayo.R
\name{egfr_mayo}
\alias{egfr_mayo}
\title{Estimated GFR Using the Mayo Formula}
\usage{
\code{egfr_mayo(..., scr.mg.dL, age.yr, is.female, scr.umol.L, scr.conversion.factor = 88.4)}
}
\arguments{
\item{...}{Ignored. This forces all parameters to be named, which is safer.}

\item{scr.mg.dL}{Numeric. Serum creatinine in mg/dL.}

\item{age.yr}{Numeric. Age in years.}

\item{is.female}{Logical. TRUE for females, FALSE for males.}

\item{scr.umol.L}{Numeric. Alternative specification of serum creatinine in umol/L.}

\item{scr.conversion.factor}{Numeric. A factor for converting serum creatinine from umol/L to mg/dL.}
}
\description{
Computes estimated glomerular filtration rate (GFR) using the Mayo Clinic quadratic formula.
}
\details{
All arguments can be vectors.
}
\examples{
# With serum creatinine in mg/dL
egfr_mayo(scr.mg.dL=0.9, age.yr=53, is.female=FALSE)

# With serum creatinine in umol/L
egfr_mayo(scr.umol.L=0.9*88.4, age.yr=53, is.female=FALSE)

}
\references{
\url{http://en.wikipedia.org/wiki/Creatinine_clearance}
}
\seealso{
\link{crcl.cg}
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
