% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_adult_vols.R
\name{est_adult_vols}
\alias{est_adult_vols}
\title{Estimate Adult Fluid Volumes}
\usage{
\code{est_adult_vols(..., wt.kg, wt.lb, is.female, return.df)}
}
\arguments{
\item{...}{Ignored. This forces all parameters to be named, which is safer.}

\item{wt.kg}{Numeric. Weight in kilograms.}

\item{wt.lb}{Numeric. Weight in pounds.}

\item{is.female}{Logical. TRUE for females, FALSE for males.}

\item{return.df}{Logical. TRUE to return \code{data.frame}, FALSE to return \code{list}.}
}
\description{
Computes estimated total body water, extracellular and intracellular fluid volumes for adults.
}
\details{
All arguments can be vectors. By default, a \code{list} is returned, though a
         \code{data.frame} may be requested instead.

         The formula used is simple, and derived from human physiology references:

         Total body water (tbw) = 0.5 * wt.kg (females), 0.6 * wt.kg (males)
         Extracellular (ECC) volume (litres) = tbw * (1/3)
         Intracellular (ICF) volume (litres) = tbw * (2/3)
         Interstitial (ISF) volume (litres) = tbw * (3/4)
         Plasma volume (litres) = tbw * (1/4)
}
\examples{
# For a male weighing 65kg, estimated total body water is 39L, ECC volume is 13L
est_adult_vols(wt.kg=65, is.female=FALSE)

# For a female weighing 50kg, estimated total body water is 25L, ECC volume is 8.33L
est_adult_vols(wt.kg=50, is.female=TRUE)

}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
