% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_obs.R
\name{expand_obs}
\alias{expand_obs}
\title{Create Template Data Sets for mrgsolve, Allowing for Varying Sampling Intervals}
\usage{
expand_obs(..., indata, dense, units, duration, sampling.int, sampling.freq)
}
\arguments{
\item{...}{passed to \code{\link{expand.grid}}}

\item{indata}{input data set}

\item{dense}{Boolean; if \code{TRUE}, sample more intensively around dose times. Otherwise, use consistent intervals only.}

\item{units}{Character string specifying units in minutes, hours, days, weeks, or years.}

\item{duration}{Numeric value specifying maximum length of simulations as specified in \code{units}.}

\item{sampling.int}{Numeric value specifying width of sampling interval around dose times in \code{units}.}

\item{sampling.freq}{Numeric value specifying sampling frequency for \code{mrgsim} in \code{units}}

\item{cmt}{Numeric value specifying compartment in which to dose; default is 1.}
}
\value{
A data frame ordered by subject, and then by time.
}
\description{
This function shares many features with \code{expand.ev} in the mrgsolve package,
but offers additional features.
}
\details{
For intensive sampling around dose times, the \code{dense} option may be requested.
The user has control over the sampling frequency (times per hour),
as well as the width of the sampling window (time before/after dosing). \code{expand_obs}
calls the \code{sample_window} function to generate the additional observations before
or after dose times.
The resulting dataset includes the number of days, as well as the time of
day for the observation, as calculated from the start of the simulated

Currently, the function assumes the dose variable is called \code{DOSE}, and the
time variable is called \code{TIME}.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
