% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2n.R
\name{f2n}
\alias{f2n}
\title{Convert Factor to Numeric}
\usage{
f2n(x)
}
\arguments{
\item{x}{A factor containing numeric information to convert to a vector of numeric values (which is no longer a factor).}
}
\description{
This function will convert a factor to a vector of numeric values. This is a simple convenience function.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
