% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_interval.R
\name{in_interval}
\alias{in_interval}
\title{Determine Whether Numeric Value Lies Within Specified Interval}
\usage{
in_interval(x, lo, hi, closed.right=TRUE, closed.left=TRUE)
}
\arguments{
\item{x}{A numeric value.}

\item{lo}{A numeric value specifying the lower bound of the interval.}

\item{hi}{A numeric value specifying the upper bound of the interval.}

\item{closed.right}{Boolean; is the interval closed on the right (upper) side? Defaults to \code{TRUE}.}

\item{closed.left}{Boolean; is the interval closed on the left (lower) side? Defaults to \code{TRUE}.}
}
\value{
\code{TRUE} if the value \code{x} lies within the interval, \code{FALSE} if it does not.
}
\description{
This is a simple function to check whether a value is contained within the interval \code{[lo, hi]}. The
function also allows specifying whether the interval is closed on the left, i.e. \code{(lo, hi])}, on the right only,
i.e. \code{[lo, hi)}, or both \code{(lo, hi)}.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
