% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interleave.R
\name{interleave}
\alias{interleave}
\title{Interleave Two Vectors}
\usage{
interleave(a, b)
}
\arguments{
\item{a}{A vector \code{A}.}

\item{b}{A second vector \code{B}.}
}
\value{
An interleaved vector with length equal to \code{length(A) + length(B)}.
}
\description{
This function takes two vectors and merges them, by means of an in shuffle. Every other value in the result vector originates
from the same source vector; values in odd-numbered positions are from vector A, while even-numbered positions are occupied
by values from vector B. The ordering of the original values is preserved.
}
\examples{
a <- c(1, 2, 3)
b <- c(4, 5, 6)

interleave(a, b) # returns 1 4 2 5 3 6
interleave(as.character(a), as.character(b)) # returns "1" "4" "2" "5" "3" "6"

}
\references{
\url{https://stackoverflow.com/questions/16443260/interleave-lists-in-r}
}
\author{
Arun Srinivasan
}
\keyword{utilities}
