% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanRange.R
\name{meanRange}
\alias{meanRange}
\title{Return Estimated Mean and Range as Character String}
\usage{
meanRange(x, digits, ssize=TRUE, type="arithmetic")
}
\arguments{
\item{x}{A vector of numeric values.}

\item{digits}{The number of significant digits to report. The default is to report three digits.}

\item{ssize}{Boolean; should the number of observations from which mean and range are computed be returned?}

\item{type}{A character string corresponding to the type of mean to be estimated, i.e. \code{arithmetic}, \code{geometric}, or \code{harmonic}.}
}
\value{
A named numeric vector, with one element if \code{ssize} is \code{FALSE}, and two elements if it is \code{TRUE}.
}
\description{
This is a very simple function, which allows for estimation of means and ranges, and formatting the result
as a character string. Significant digits may be specified, and the number of values from which the mean and range are computed
may be returned if \code{ssize} is \code{TRUE}.
}
\examples{

}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
