% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_ws.R
\name{rm_ws}
\alias{rm_ws}
\title{Completely or Selectively Clear the R Workspace of Objects}
\usage{
rm.ws(..., exclude, verbose=FALSE, reqmatch=TRUE)
}
\arguments{
\item{...}{Arguments to be passed directly to \code{rm}.}

\item{exclude}{Boolean; should any objects be retained when the workspace is cleared?}

\item{verbose}{Boolean; should output be written to console listing the deleted/retained objects?}

\item{reqmatch}{Boolean; should \code{rm.ws} require that at least one exclusion matches?}
}
\value{
None
}
\description{
This is a convenience function for cleaning the R workspace; either all objects can be removed, or only a subset of them.
}
\details{
If exclusions are specified, but none are matched, the function will abort for safety reasons. This is intended to prevent
the user from inadvertently clearing the workspace using stale or poorly formatted arguments.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
