% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_window.R
\name{sample_window}
\alias{sample_window}
\title{Generate Intensive Sampling Times for mrgsolve Conditional on Dose Times}
\usage{
sample_window(dosetime, interval, sampfreq, include = FALSE,
  before = FALSE, negtime = FALSE)
}
\arguments{
\item{dosetime}{Numeric. Vector of dose administration times.}

\item{interval}{Numeric. A single value corresponding to the width of the sampling interval.}

\item{sampfreq}{Numeric. The frequency of sampling (in fractions of an hour).}

\item{include}{Logical. Should dose times be included for sampling? Default is to avoid resampling at dose times.}

\item{before}{Logical. Should additional samples be taken before dose times, or only after a dose is administered?}

\item{negtime}{Logical. Should negative times be allowed? Default is to disallow negative observation times.}
}
\description{
The \code{sample_window} function uses dose administration times to generate additional observations for insertion
into a mrgsolve dataset. The more intensive sampling around drug doses may help more accurately estimate quantities
such as AUC, while avoiding potentially time-consuming oversampling at other times.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
