% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortlevels.R
\name{sortlevels}
\alias{sortlevels}
\title{Sort and Format Levels for Factoring}
\usage{
sortlevels(all.visits, pre.levels, excluded.levels, title.case=TRUE)
}
\arguments{
\item{all.visits}{A vector of all visit names to sort and format.}

\item{pre.levels}{A vector of visit levels that should precede the rest (e.g. screening or pre-study visits) in ordering.}

\item{excluded.levels}{A vector of visit names that should be excluded from the factor levels.}

\item{title.case}{Logical. Should the names of the factor levels use title case?}
}
\value{
A vector of visit levels, sorted and formatted, potentially with exclusions.
}
\description{
This function will reformat and order visit names for use in factors; it also allows modifying case and excluding some levels.
}
\author{
Ben Rich, \email{benjamin.rich@certara.com} with edits by Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
