% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_verbose.R
\name{source_verbose}
\alias{source_verbose}
\title{Verbosely Read R Code from a File, a Connection, or Expressions}
\usage{
source_verbose(..., envir = .GlobalEnv)
}
\arguments{
\item{...}{Arguments to pass to \code{source}.}
}
\description{
\code{source_verbose} behaves identically to \code{\link{source}}, but prints a list of objects added to or removed from
the chosen environment to the console. The function is a simple wrapper; if the workspace is unchanged, nothing is printed.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
