% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepart_dtc.R
\name{timepart_dtc}
\alias{timepart_dtc}
\title{Extract Time From Date/Time Objects}
\usage{
timepart_dtc(x, format, outfmt, tz)
}
\arguments{
\item{x}{A date-time object, as a character string, formatted according to \code{format}.}

\item{format}{The format of \code{x}; default is to assume that a \code{T} separates the date from the time, as with CDISC data.}

\item{outfmt}{The format (as character string) to use for the result.}

\item{tz}{The desired \link{timezone}; default is "\code{UTC}".}
}
\value{
A character string with the time portion of \code{x}.
}
\description{
This function takes a date-time object and its format as inputs, and returns the time portion.
}
\author{
Ryan Kyle, \email{ryan.kyle@mail.mcgill.ca}
}
\keyword{utilities}
