% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTitleCase_always.R
\name{toTitleCase_always}
\alias{toTitleCase_always}
\title{Always Convert Text to Title Case}
\usage{
toTitleCase_always(text)
}
\arguments{
\item{text}{A character string to be converted to title case.}
}
\value{
A character vector of the same length as \code{text}, without names.
}
\description{
Convert a character vector to title case, less carefully than \code{toTitleCase} does. Unlike \code{toTitleCase}, requires loading no packages (other than this one).
}
\details{
Modified by Ryan Kyle, based on function from \url{https://stackoverflow.com/questions/6364783/capitalize-the-first-letter-of-both-words-in-a-two-word-string}.
}
\author{
Andrie de Vries
}
\keyword{utilities}
