% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecseq.R
\name{vecseq}
\alias{vecseq}
\title{Vectorized Sequence Generation}
\usage{
seq(from = 1, to = 1, by = ((to - from)/(length.out - 1)), length.out = NULL, along.with = NULL, ...)
}
\arguments{
\item{from}{the starting value of the sequence. Of length 1 unless just from is supplied as an unnamed argument.}

\item{to}{the ending (maximal) value of the sequence. Of length 1 unless just from is supplied as an unnamed argument.}

\item{by}{number: increment of the sequence.}

\item{length.out}{desired length of the sequence. A non-negative number, which for seq and seq.int will be rounded up if fractional.}

\item{along.with}{take the length from the length of this argument.}

\item{...}{arguments passed to or from methods.}
}
\value{
A matrix of vectors, with each column corresponding to a sequence specified by \code{from} and \code{to}.
}
\description{
Generate regular sequences. \code{seq} is a standard generic with a default method. This is a vectorized version of the \code{seq}
function from base R.
}
\examples{
vecseq(from = c(1, 1.75), to = c(2, 2.75), by = 0.25)

# results in:
#      [,1] [,2]
# [1,] 1.00 1.75
# [2,] 1.25 2.00
# [3,] 1.50 2.25
# [4,] 1.75 2.50
# [5,] 2.00 2.75

}
\author{
Magically generated by the \code{\link{Vectorize}} function.
}
\keyword{utilities}
